% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_epochr.R
\name{ag_epochr}
\alias{ag_epochr}
\title{Function to re-integrate 1-second actigraph count data into a larger epoch}
\usage{
ag_epochr(ag_data_1sec, epoch = 60)
}
\arguments{
\item{ag_data_1sec}{ActiGraph count data exported in 1-second format}

\item{epoch}{Numeric integer for the desired reintegrated epoch length}
}
\description{
Function to re-integrate 1-second actigraph count data into a larger epoch
}
