% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_feature_calc.R
\name{ag_feature_calc}
\alias{ag_feature_calc}
\title{Custom function to compute features of raw accelerometer data. Can be applied to fixed- or variable-windows (epochs vs sojourns, respectively).}
\usage{
ag_feature_calc(
  ag_data_raw_wrist,
  participant,
  samp_freq = 80,
  window = 15,
  soj_colname = NA,
  seconds_colname = NA,
  inactive_threshold = 0.1
)
}
\arguments{
\item{ag_data_raw_wrist}{Raw accelerometer data containing AxisX, AxisY, AxisZ, and VM data columns}

\item{participant}{Character string for participant ID}

\item{samp_freq}{Sampling frequency of the raw accelerometer data. Defaults to 80 Hz}

\item{window}{Epoch window (in seconds). Can be either a fixed number or 'sojourn' character string. If 'sojourn', a column of sojourn indices is expected to be present in the data.}

\item{soj_colname}{Character string for column name containing the sojourn indices}

\item{seconds_colname}{Character string for column name containing the duration of a given sojourn
Library dependencies: tidyverse, lubridate, data.table}
}
\description{
Custom function to compute features of raw accelerometer data. Can be applied to fixed- or variable-windows (epochs vs sojourns, respectively).
}
