% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion_intensity_AP_HR.R
\name{criterion_intensity_AP_HR}
\alias{criterion_intensity_AP_HR}
\title{Function to create criterion measure from activpal and heart rate data}
\usage{
criterion_intensity_AP_HR(HR_data, AP_data, HR_flex, MPA_HRR, VPA_HRR)
}
\arguments{
\item{HR_data}{Data with a Timestamp in ymd_hms format and heart rate data labelled as 'HR'}

\item{AP_data}{ActivPAL data with a Timestamp column in ymd_hms format and event data (sitting/lying, standing, stepping)}

\item{HR_flex}{Individualized HR-flex threshold. Periods with HR data above this threshold will use HR to determine intensity while periods below will use activPAL event data}

\item{MPA_HRR}{Individualized threshold for determining moderate-intensity activity from HR data}

\item{VPA_HRR}{Individualized threshold for determining vigorous-intensity activity from HR data}
}
\description{
Function to create criterion measure from activpal and heart rate data
}
