% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dev_DO_Viz.R
\name{Dev_DO_Viz}
\alias{Dev_DO_Viz}
\title{Function to visualize device data with direct observation data that have been generated by AG_DO_merge function
Often used to visually screen for any anomalies or temporal misalignment between device and direct observation data}
\usage{
Dev_DO_Viz(rds_filepaths, output_filepath, overwrite = T, samp_freq = 80)
}
\arguments{
\item{overwrite}{= T}

\item{samp_freq}{= 80

Library dependencies: ggplot2, dplyr, tidyr, stringr}
}
\description{
Function to visualize device data with direct observation data that have been generated by AG_DO_merge function
Often used to visually screen for any anomalies or temporal misalignment between device and direct observation data
}
