% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ag.R
\name{read_ag}
\alias{read_ag}
\title{Function to read in ActiGraph accelerometer data.
Filepath provided should be to an excel .csv format file
Accelerometer data should be exported with steps, lux, and inclinometer data. Otherwise, may result in improper column naming}
\usage{
read_ag(
  filepath,
  ENMO_calibrate = T,
  device_serial_calibrate = T,
  calibration_file,
  parse_timestamp = T
)
}
\arguments{
\item{filepath}{Character string of complete filepath for a single ActiGraph accelerometer excel file. May be raw or count-based data}

\item{ENMO_calibrate}{Binary indicator (T/F) of whether ENMO calibrated acceleration data should be computed. Uses g.calibrate function from GGIR (broken with recent GGIR updates)}

\item{device_serial_calibrate}{Binary indicator (T/F) of whether to use device-specific correction factor coefficients when there is insufficient non-movement periods in the data file for auto-calibration}

\item{calibration_file}{Data frame containing device serial numbers and device-specific correction factor coefficients}

\item{parse_timestamp}{Binary indicator (T/F) of whether the Timestamp should be separated into Date and Time columns. For larger files, this extends the read and return time.
Library dependencies: GGIR, stringr, data.table, dplyr}
}
\description{
Function to read in ActiGraph accelerometer data.
Filepath provided should be to an excel .csv format file
Accelerometer data should be exported with steps, lux, and inclinometer data. Otherwise, may result in improper column naming
}
