% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ap.R
\name{read_ap}
\alias{read_ap}
\title{Function to read in activPAL data}
\usage{
read_ap(
  activpal_filepath,
  raw_data = F,
  parse_timestamp = F,
  epoch = NA,
  samp_freq = 20
)
}
\arguments{
\item{activpal_filepath}{A .csv filepath containing either event data processed by the activPAL algorithm or raw accelerometer data}

\item{raw_data}{A logical indicating whether the data is raw data. Default is F and event data is expected}

\item{parse_timestamp}{A logical indicating whether the Timestamp in ymd_hms format should be further separated into Date and Time columns}

\item{epoch}{}
}
\description{
Function to read in activPAL data
}
