% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedsphere.R
\name{sedsphere}
\alias{sedsphere}
\title{Function to apply the Sedentary Sphere from Rowlands et al 2014}
\usage{
sedsphere(
  acc_data_raw,
  VMcorrG_mod_15s = 489,
  samp_freq = 80,
  epoch = 15,
  expand_1sec = F
)
}
\arguments{
\item{acc_data_raw}{Raw tri-axial wrist accelerometer data with column names Timestamp, AxisX, AxisY, AxisZ}

\item{VMcorrG_mod_15s}{Moderate-to-vigorous intensity activity acceleration threshold. Default MVPA threshold is >489 g}

\item{samp_freq}{Sampling frequency of the raw accelerometer data. Default is 80 hz}

\item{epoch}{Non-overlapping window size in seconds. Default is 15-seconds}

\item{expand_1sec}{Binary indicator of whether only SedSphere estimates should be returned as a second-by-second vector}
}
\value{
Aggregated data in 15-second epochs with accelerometer values and SedSphere estimate
}
\description{
Function to apply the Sedentary Sphere from Rowlands et al 2014
}
