% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soj_g.R
\name{soj_g}
\alias{soj_g}
\title{Soj-g function from Marcotte et al 2021}
\usage{
soj_g(
  data = NA,
  export_format = "session",
  freq = 80,
  step1_sd_threshold = 0.00375,
  step2_nest_length = 5,
  step3_nest_length = 60,
  step3_orig_soj_length_min = 180
)
}
\arguments{
\item{data}{A dataframe with raw wrist accelerometer data with column names Timestamp, AxisX, AxisY, AxisZ, VM}

\item{freq}{Sampling frequency of raw accelerometer data. 80 Hz by default}

\item{step1_sd_threshold}{Threshold used to classify likely inactive periods}

\item{step2_nest_length}{Window length for classifying inactive vs active using random forest model}

\item{step3_nest_length}{Window length for partitioning excessively long sojourns as determined by step3_orig_soj_length_min}

\item{step3_orig_soj_length_min}{Maximum sojourn window duration allowed until partitioning into nested sojourns}
}
\description{
Soj-g function from Marcotte et al 2021
}
